refactor: replace isr hardcoding with isr_table, add ssize_t, refine type casting

+++ include/string.h
@@ -3,6 +3,6 @@
 
 #include "types.h"
 
-i32 strlen(const char* str);
+ssize_t strlen(const char* str);
 
 #endif
